Compilation fix
git-svn-id: file:///home/roberto/Git/svn@10931 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 1cd19004886ab71254de2e6423ff87a074322e82

+++ trunk/xmipp/libraries/data/multidim_array.h
@@ -2444,9 +2444,9 @@ public:
         y -= STARTINGY(*this);
         x -= STARTINGX(*this);
 
-        int l1 = ceil(x - SplineDegree_1);
+        int l1 = (int)ceil(x - SplineDegree_1);
         int l2 = l1 + SplineDegree;
-        int m1 = ceil(y - SplineDegree_1);
+        int m1 = (int)ceil(y - SplineDegree_1);
         int m2 = m1 + SplineDegree;
 
         double columns = 0.0;
